Specify multiple params to an action.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@3692 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qs/src/util/commandline.cpp
@@ -84,7 +84,7 @@ bool qs::CommandLine::parse(Reader* pReader)
 			if (!bPrevSpace) {
 				if (!pImpl_->pHandler_->process(buffer.getCharArray()))
 					return false;
-				buffer.remove(0, static_cast<size_t>(-1));
+				buffer.remove(0, -1);
 			}
 			bSpace = true;
 		}
